ats_utils/extract_docx.py
from docx import Document

def extract_text_from_docx(file_path: str) -> str:
    try:
        doc = Document(file_path)
        text = "\n".join([para.text for para in doc.paragraphs])
        return text.strip()
    except Exception as e:
        return f"[ERROR extracting DOCX] {str(e)}"
