requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
numpy
pdfminer.six
langchain
ollama
python-docx
python-dotenv requests
PyMuPDF
requests
firebase-admin
